JPA BookDao and BookDaoTests

+++ library/src/main/java/ru/otus/spring01/library/dao/BookDao.java
@@ -2,13 +2,12 @@ package ru.otus.spring01.library.dao;
 
 import ru.otus.spring01.library.domain.Book;
 import ru.otus.spring01.library.domain.Genre;
-import ru.otus.spring01.library.domain.Person;
 
 import java.util.List;
 import java.util.UUID;
 
 public interface BookDao {
-    int count();
+    Long count();
 
     void insert(Book book);
